Fix some token state machine bugs.
Change the em-dash handling to only emit the word and dash once the run
of dashes ends, rather than always emitting after two dashes.

Jump directly to the running dash state when a dash is encountered in
the whitespace state.

Use only the match-all input predicate where redundant with the
whitespace predicate (so that unknown characters are treated as
whitespace).

Add a few more tests to make sure that punctuation after words and
longer em-dashes are correctly tokenized.